# judge/openapi.yaml
openapi: 3.1.0
info:
  title: Judge Service API
  version: "1.0.0"
  description: |
    Code judging for the auction-based coding competition.

    This service answers exactly one question: **does this source code pass this
    problem?** It knows nothing about auctions, balances, question ownership,
    hints or scoring — all of that belongs to the backend.

    ### Submit, then poll
    `POST /submit` returns `202` with a job handle; poll `GET /jobs/{job_id}`
    until `state` is `done`. Judging takes anywhere from under a second to well
    over a minute, depending on how close the solution runs to its time limit.

    Job state is held **in memory only**. Nothing is persisted — keep
    submissions in your own database.

    ### Deployment
    Runs on the internal network only. It fronts `go-judge`, which executes
    untrusted code in a sandbox and must never be exposed directly.

    ### Leaking answers
    `Judgement.jury_detail` contains the expected output for the failing test.
    Never render it to a contestant. See the field description.
servers:
  - url: http://judge-api:8000
    description: Internal Docker network
security:
  - bearerAuth: []
...
paths:
...
  /languages:
    get:
      tags: [Metadata]
      summary: Supported languages
      description: |
        Drive the submission language dropdown from this endpoint so the UI and
        the judge cannot drift apart.

        Every language gets the **same** time limit, as on Codeforces — there are
        no per-language multipliers. Interpreted languages are therefore at a
        real disadvantage on tight problems; set limits generously relative to
        the intended solution, and say so in the contest rules so participants
        can factor it into their bidding.
      responses:
        "200":
          description: Available languages
          content:
            application/json:
              schema:
                type: object
                required: [languages]
                properties:
                  languages:
                    type: array
                    items: { $ref: "#/components/schemas/Language" }
              example:
                languages:
                  - { key: c,          name: "C (GCC 14)",          compiled: true }
                  - { key: cpp,        name: "C++ (GCC 14)",        compiled: true }
                  - { key: python,     name: "Python 3.13",         compiled: false }
                  - { key: pypy,       name: "PyPy 3.11 (7.3.23)",  compiled: false }
                  - { key: java,       name: "Java 21",             compiled: true }
                  - { key: javascript, name: "JavaScript (Node 20)", compiled: false }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /problems/{problem_id}:
    get:
      tags: [Metadata]
      summary: Problem metadata
      description: |
        Call this at **auction setup** to confirm every question you are about to
        sell actually exists and has testcases. Discovering a missing problem
        after a participant has paid for it is unrecoverable mid-contest.
      parameters:
        - name: problem_id
          in: path
          required: true
          schema: { type: string, pattern: "^[A-Za-z0-9._-]+$" }
          example: hard-03
      responses:
        "200":
          description: Problem metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemInfo" }
              example:
                problem_id: hard-03
                testcases: 87
                time_limit_ms: 2000
                memory_limit_mb: 256
                compare: checker
                early_exit: true
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProblemNotFound" }
  /problems:
    get:
      tags: [Metadata]
      summary: List all problems
      operationId: listProblems
      description: |
        Everything the admin dashboard needs to show the state of the problem
        set: counts, sizes, versions and whether each one has been validated.
        This is the "file manager" view — a page in your admin UI, not a
        separate service.
      responses:
        "200":
          description: All problems known to the judge.
          content:
            application/json:
              schema:
                type: object
                required: [problems]
                properties:
                  problems:
                    type: array
                    items: { $ref: "#/components/schemas/ProblemInfo" }
              example:
                problems:
                  - { problem_id: easy-01, testcases: 24, bytes: 8310422, version: v2,
                      time_limit_ms: 1000, memory_limit_mb: 256, compare: tokens,
                      early_exit: true, validated: true, modified_at: "2026-09-09T14:02:11Z" }
                  - { problem_id: hard-03, testcases: 87, bytes: 149002133, version: v3,
                      time_limit_ms: 1000, memory_limit_mb: 256, compare: checker,
                      early_exit: true, validated: false, modified_at: "2026-09-10T09:41:55Z" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /problems/{problem_id}/validate:
    post:
      tags: [Metadata]
      summary: Validate a problem before the contest
      operationId: validateProblem
      description: |
        Checks the problem is actually usable: every `.in` has a matching answer
        file, `problem.json` parses, and the checker compiles.

        If a `reference_source` is supplied it is run against **every** testcase.
        This is the part that matters — it proves the answer files are correct.
        A `reference.first_fail` of 61 means testcase 61's answer file is wrong,
        found on a Tuesday rather than during the contest.

        Run it for all problems as the last step of deployment, and wire it to a
        "Validate" button in the admin UI.
      parameters:
        - name: problem_id
          in: path
          required: true
          schema: { type: string, pattern: "^[A-Za-z0-9._-]+$" }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reference_source:
                  type: string
                  description: A known-correct solution. Omit to run structural checks only.
                language: { type: string, example: cpp }
                wrong_source:
                  type: string
                  description: |
                    Optionally, a known-*incorrect* solution. If this is
                    accepted, the testcases fail to catch a bad answer — a test
                    set that passes everything is worse than useless.
      responses:
        "200":
          description: "Validation report. A false `ok` is a normal result, not an error."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationReport" }
              examples:
                bad_answer_file:
                  summary: Reference solution fails — an answer file is wrong
                  value:
                    problem_id: hard-03
                    ok: false
                    testcases: 87
                    version: v3
                    issues: []
                    checker: { compiled: true }
                    reference: { verdict: WA, first_fail: 61, passed: 60 }
                    wrong_solution: null
                structural:
                  summary: Missing answer file
                  value:
                    problem_id: hard-03
                    ok: false
                    testcases: 86
                    version: v3
                    issues: ["00042.in has no matching answer file"]
                    checker: { compiled: true }
                    reference: null
                    wrong_solution: null
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProblemNotFound" }
  /problems/{problem_id}/testcases/{index}:
    get:
      tags: [Metadata]
      summary: Read one testcase (for paid hint reveals)
      operationId: getTestcase
      description: |
        Returns a testcase's input and expected answer.

        **This exists so your backend can sell a testcase as a paid hint.** The
        judge does not know or care who paid — it serves bytes to an
        authenticated caller. Access control is entirely your backend's job.

        **Always pass `version`**, taken from the submission's stored
        `problem_version`. Without it you get whatever is current, which after a
        mid-contest edit is not the testcase their submission failed on — and
        they paid for that specific one.

        Never expose this endpoint to contestants directly.
      parameters:
        - name: problem_id
          in: path
          required: true
          schema: { type: string, pattern: "^[A-Za-z0-9._-]+$" }
        - name: index
          in: path
          required: true
          description: Zero-based, matching `first_fail` in a judgement.
          schema: { type: integer, minimum: 0 }
          example: 12
        - name: version
          in: query
          required: false
          description: Problem version, from the submission's `problem_version`.
          schema: { type: string }
          example: v3
      responses:
        "200":
          description: The testcase.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Testcase" }
              example:
                problem_id: hard-03
                index: 12
                version: v3
                input: "3\n1 2\n5 7\n10 20\n"
                answer: "3\n12\n30\n"
                truncated: false
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProblemNotFound" }
  /submit:
    post:
      tags: [Judging]
      summary: Queue a submission for judging
      operationId: submit
      description: |
        Accepts a submission and returns immediately with a job handle. Poll
        `GET /jobs/{job_id}` for progress and the final verdict.

        **Why not synchronous:** judging time scales with the solution's own
        runtime. A correct solution that sits just under a 2s limit across ~100
        testcases takes ~38s, and a Python one with its 3x multiplier can exceed
        100s. Those are exactly the submissions that deserve to pass, and they
        are the ones a blocking call would lose to a proxy timeout.

        The judge stores nothing durably. Keep the source in your own database
        so a rejudge — or recovery after a judge restart — is just resubmitting.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitRequest" }
            examples:
              cpp:
                value:
                  problem_id: hard-03
                  language: cpp
                  source: "#include <iostream>\nint main(){int t;std::cin>>t;while(t--){long long a,b;std::cin>>a>>b;std::cout<<a+b<<'\\n';}}"
                  submission_id: sub_8f21c
      responses:
        "202":
          description: Accepted for judging. Poll the job for the result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobHandle" }
              example:
                job_id: job_01JQ8XKZ3M
                submission_id: sub_8f21c
                state: queued
                poll_after_ms: 500
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProblemNotFound" }
        "429": { $ref: "#/components/responses/Busy" }
        "503": { $ref: "#/components/responses/SandboxDown" }
  /hack:
    post:
      tags: [Judging]
      summary: Does this input break the given solution?
      operationId: hack
      description: |
        Phase 1's Hacking section. The participant reads a deliberately flawed
        solution and submits a **test input** they believe breaks it.

        Three runs, in order: the problem's validator on the input (an invalid
        input is not a hack), the problem's **stored reference solution** to get
        the correct answer, then the given solution compared against it. It is
        `hacked` if the given solution gets a wrong answer, exceeds a limit, or
        crashes.

        The problem must carry a reference solution (`problem.json` →
        `reference`). Anything wrong with the problem itself — reference missing
        or failing, given solution not compiling — is `IE`, and `hacked` is
        `null`: the participant is neither credited nor penalised.

        **Never show a participant `verdict`.** Whether it was a wrong answer or
        a timeout tells them the shape of the bug; that turns the judge into an
        oracle for probing the solution instead of a check on reading it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HackRequest" }
      responses:
        "202":
          description: Accepted. Poll the job; `result` is a HackResult.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobHandle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ProblemNotFound" }
        "429": { $ref: "#/components/responses/Busy" }
        "503": { $ref: "#/components/responses/SandboxDown" }
  /validate-answers:
    post:
      tags: [Judging]
      summary: Score a list of answers with a supplied validator
      operationId: validateAnswers
      description: |
        Phase 1's validator-scored questions — "find as many valid passwords as
        you can" — where there is no fixed answer key. The validator is a Python
        file defining `check(entry)`; each entry is a Reader over one submitted
        answer, using the same API as checkers.

        Runs in the sandbox under its own limits, because a typo in an
        administrator's validator must not be able to hang the contest.

        If the validator itself cannot run, `status` is `IE` and `results` is
        empty. **Treat that as "nothing was checked"** — never as every entry
        being invalid, or one bad validator zeroes every participant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnswersRequest" }
      responses:
        "202":
          description: Accepted. Poll the job; `result` is an AnswersResult.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobHandle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/Busy" }
        "503": { $ref: "#/components/responses/SandboxDown" }
  /jobs/{job_id}:
    get:
      tags: [Judging]
      summary: Poll a judging job
      operationId: getJob
      description: |
        Returns the job's current state. Poll roughly once a second; the
        response carries `poll_after_ms` as a hint.

        `state` is `queued`, `running`, or `done`. Only when it is `done` is
        `result` present — and a failing verdict is still a completed job, not
        an error.

        **Jobs are held in memory** with a TTL (default 10 minutes after
        completion). A restarted judge loses in-flight jobs and this returns
        `404`; resubmit from your stored source.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
          example: job_01JQ8XKZ3M
      responses:
        "200":
          description: Current job state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobState" }
              examples:
                queued:
                  summary: Waiting for a judging slot
                  value:
                    job_id: job_01JQ8XKZ3M
                    submission_id: sub_8f21c
                    state: queued
                    progress: { done: 0, total: 87 }
                    poll_after_ms: 500
                    result: null
                running:
                  summary: Compiled, now running testcases
                  value:
                    job_id: job_01JQ8XKZ3M
                    submission_id: sub_8f21c
                    state: running
                    progress: { done: 34, total: 87 }
                    poll_after_ms: 1000
                    result: null
                done_accepted:
                  summary: Finished — accepted
                  value:
                    job_id: job_01JQ8XKZ3M
                    submission_id: sub_8f21c
                    state: done
                    progress: { done: 87, total: 87 }
                    poll_after_ms: null
                    result:
                      submission_id: sub_8f21c
                      verdict: AC
                      passed: 87
                      total: 87
                      first_fail: null
                      max_time_ms: 412.3
                      max_memory_kb: 34120
                      compile_output: ""
                      message: all testcases passed
                      jury_detail: ""
                      duration_ms: 3310
                done_wrong_answer:
                  summary: Finished — wrong answer, stopped early
                  value:
                    job_id: job_01JQ8XKZ3N
                    submission_id: sub_8f21d
                    state: done
                    progress: { done: 12, total: 87 }
                    poll_after_ms: null
                    result:
                      submission_id: sub_8f21d
                      verdict: WA
                      passed: 11
                      total: 87
                      first_fail: 12
                      max_time_ms: 88.1
                      max_memory_kb: 12040
                      compile_output: ""
                      message: wrong answer on test 12
                      jury_detail: "token 2: expected '41', got '39'"
                      duration_ms: 402
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/JobNotFound" }
    delete:
      tags: [Judging]
      summary: Cancel a job
      operationId: cancelJob
      description: |
        Cancels a queued or running job. Use it when a participant resubmits
        before the previous judgement finished — otherwise a few people
        resubmitting during a slow problem can saturate every judging slot with
        results nobody will read.

        Cancelling an already-finished job is a no-op.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: "Cancelled, or already finished." }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/JobNotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared secret between the backend and the judge.
  schemas:
...
    Language:
      type: object
      required: [key, name]
      properties:
        key: { type: string, example: cpp }
        name: { type: string, example: "C++ (GCC 14)" }
        compiled:
          type: boolean
          description: Whether the language has a compile step (affects CE verdicts).
          example: true
...
    Testcase:
      type: object
      required: [problem_id, index, input, answer]
      properties:
        problem_id: { type: string }
        index: { type: integer, example: 12 }
        version: { type: string, example: v3 }
        input: { type: string }
        answer: { type: string }
        truncated:
          type: boolean
          description: True if the testcase exceeded the response size cap and was cut.
...
  responses:
...
    Unauthorized:
      description: Missing or bad bearer token. Treat as a configuration bug, not user error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: unauthorized, message: invalid or missing bearer token }
    ProblemNotFound:
      description: |
        No such problem. During a contest this is serious — it means a question
        being auctioned has no testcases behind it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: problem_not_found, message: "no such problem: hard-99" }
...